+++ Programas/ejercicioSupermercado/src/com/foc/interfaces/EsAlimento.java
@@ -5,5 +5,5 @@ import java.time.LocalDate;
 public interface EsAlimento {
     public void setCaducidad(LocalDate fecha);
     public LocalDate getCaducidad();
-    public int getCalorias();
+    public double getCalorias();
 }
